Property cache now uses paths, is exclusive with RDG
The cache takes a simple katana::Uri as its key, which
allows a straightforward and generic cache implementation
without concepts specific to property caches.

Also, the property cache is now exclusive with
the graph, meaning a property column either resides
in the graph or in the cache, never both.  This makes
the cache more like fast storage, it can always be cleared,
and we never evict an in-use property.